chore: move deleteSavedAddress mutation to folder
