+++ src/lib/Scenes/SavedAddresses/SavedAddresses.tsx
@@ -1,6 +1,5 @@
 import { captureMessage } from "@sentry/react-native"
 import { SavedAddresses_me } from "__generated__/SavedAddresses_me.graphql"
-import { SavedAddressesDeleteUserAddressMutation } from "__generated__/SavedAddressesDeleteUserAddressMutation.graphql"
 import { SavedAddressesQuery } from "__generated__/SavedAddressesQuery.graphql"
 import { PageWithSimpleHeader } from "lib/Components/PageWithSimpleHeader"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
@@ -12,8 +11,9 @@ import { Button, color, Flex, Separator, Spacer, Text, Touchable } from "palette
 import React, { useCallback, useState } from "react"
 import { FlatList, RefreshControl } from "react-native"
 import { createRefetchContainer, QueryRenderer, RelayRefetchProp } from "react-relay"
-import { commitMutation, graphql } from "relay-runtime"
+import { graphql } from "relay-runtime"
 import styled from "styled-components/native"
+import { deleteUserAddress } from "./mutations/deleteSavedAddress"
 
 interface CardProps {
   isDefault: boolean
@@ -166,54 +166,6 @@ export const SavedAddressesContainer = createRefetchContainer(
     }
   `
 )
-
-const deleteUserAddress = (userAddressID: string, onSuccess: () => void, onError: (message: string) => void) => {
-  commitMutation<SavedAddressesDeleteUserAddressMutation>(defaultEnvironment, {
-    variables: {
-      input: {
-        userAddressID,
-      },
-    },
-    mutation: graphql`
-      mutation SavedAddressesDeleteUserAddressMutation($input: DeleteUserAddressInput!) {
-        deleteUserAddress(input: $input) {
-          userAddressOrErrors {
-            ... on UserAddress {
-              id
-              internalID
-              name
-              addressLine1
-              addressLine2
-              addressLine3
-              city
-              region
-              postalCode
-              phoneNumber
-              isDefault
-            }
-            ... on Errors {
-              errors {
-                code
-                message
-              }
-            }
-          }
-        }
-      }
-    `,
-    onError: (e) => {
-      onError(e.message)
-    },
-    onCompleted: (data) => {
-      const errors = data?.deleteUserAddress?.userAddressOrErrors.errors
-      if (errors) {
-        onError(errors.map((error) => error.message).join(", "))
-      } else {
-        onSuccess()
-      }
-    },
-  })
-}
 
 export const SavedAddressesQueryRenderer: React.FC<{}> = ({}) => {
   return (
