+++ src/lib/Scenes/SavedAddresses/mutations/deleteSavedAddress.ts
@@ -1,0 +1,51 @@
+import { SavedAddressesDeleteUserAddressMutation } from "__generated__/SavedAddressesDeleteUserAddressMutation.graphql"
+import { defaultEnvironment } from "lib/relay/createEnvironment"
+import { commitMutation, graphql } from "relay-runtime"
+
+export const deleteUserAddress = (userAddressID: string, onSuccess: () => void, onError: (message: string) => void) => {
+  commitMutation<SavedAddressesDeleteUserAddressMutation>(defaultEnvironment, {
+    variables: {
+      input: {
+        userAddressID,
+      },
+    },
+    mutation: graphql`
+      mutation SavedAddressesDeleteUserAddressMutation($input: DeleteUserAddressInput!) {
+        deleteUserAddress(input: $input) {
+          userAddressOrErrors {
+            ... on UserAddress {
+              id
+              internalID
+              name
+              addressLine1
+              addressLine2
+              addressLine3
+              city
+              region
+              postalCode
+              phoneNumber
+              isDefault
+            }
+            ... on Errors {
+              errors {
+                code
+                message
+              }
+            }
+          }
+        }
+      }
+    `,
+    onError: (e) => {
+      onError(e.message)
+    },
+    onCompleted: (data) => {
+      const errors = data?.deleteUserAddress?.userAddressOrErrors.errors
+      if (errors) {
+        onError(errors.map((error) => error.message).join(", "))
+      } else {
+        onSuccess()
+      }
+    },
+  })
+}
